move config-related stuff in app/token into config.go and config_test.go
